Add files via upload
This script can be used to open large-volume Zarr image arrays in Neuroglancer. It relies on Neuroglancer
and TensorStore. It only needs to be pointed at a directory containing '.zarray'. Be sure to run Python (3.8+) in
interactive mode by specifying the '-i' cli option. There are two Python modules required:  Neuroglancer, TensorStore